Support statement type configuration & performance enhancement

+++ Src/Scraper.py
@@ -1,5 +1,7 @@
 from selenium import webdriver
 from selenium.webdriver.support.ui import Select
+from pathlib import Path
+from time import sleep
 import requests
 import sys
 import threading
@@ -24,12 +26,14 @@ def clickDownload(mon_odd_row, driver):
         global done
         done = True
         href_a.click()
+    elif(type == 'a'):
+        href_a.click()
     else:
         return
 
     # switch to new open window
-    new_win = driver.window_handles[1]
-    driver.switch_to.window(new_win)
+    #new_win = driver.window_handles[1]
+    #driver.switch_to.window(new_win)
 
     # print(driver.page_source)
     pdf_file = driver.find_element_by_tag_name("iframe")
@@ -50,16 +54,17 @@ def clickDownload(mon_odd_row, driver):
         # c = {cookie['name']: cookie['value']}
         s.cookies.set(cookie['name'], cookie['value'])
 
-    local_filename = "some_path/" + download_type + "/" + filename + ".pdf"
-    r = s.get(path)
-    open(local_filename, 'wb').write(r.content)
+    # customize to yours
+    local_filename = "/yourfolder" + download_type + "/" + filename + ".pdf"
+    file = Path(local_filename)
+    if not file.is_file():
+        r = s.get(path)
+        open(local_filename, 'wb').write(r.content)
+    else:
+        print('\n' + filename + " File already existed")
 
-    driver.close()
-    driver.switch_to.window(cur_win)
-
-def download(_type, _month_year, _day ):
+def download(_type, _month_year, _day, _crdr_type ):
     # day checking
-    # note you must change the value according to your credit card statement issue day
     if _day < 15 and _type == 'd':
         print("Current Monthly credit card statments is not issued.")
         return
@@ -71,7 +76,7 @@ def download(_type, _month_year, _day ):
     type = _type
     global month_year_str
     month_year_str = _month_year
-    print('Type:' + type + ' Month-Year: ' + month_year_str)
+    print('Type:' + type + ' Month-Year: ' + month_year_str + '\n')
 
     e = threading.Event()
     t1 = threading.Thread(name='scheduler',
@@ -79,33 +84,31 @@ def download(_type, _month_year, _day ):
                           args=(e, 1))
     t1.start()
 
-    sec_pwd = ['0', 's', 'e', 'c', 'p', 'w', 'd']
+    # seccond
+    sec_pwd = []
 
     # Create a new instance of the Firefox driver
-    # driver = webdriver.Chrome("/Users/kelvinleung/PycharmProjects/chromedriver")
-    driver = webdriver.PhantomJS(executable_path=r'/path/Downloads/phantomjs-2.1.1-macosx/bin/phantomjs')
+    # driver = webdriver.Chrome("/yourpath/chromedriver")
+    driver = webdriver.PhantomJS(executable_path=r'/yourpath/phantomjs-2.1.1-macosx/bin/phantomjs')
 
     # go to the google home page
     driver.get("https://www2.ebanking.hsbc.com.hk/1/2/logon?LANGTAG=en&COUNTRYTAG=US")
 
     # find username input
     usr_name_input = driver.find_element_by_name("u_UserID")
-    usr_name_input.send_keys("your_user_id)
+    usr_name_input.send_keys("your_usr_name")
 
-    # submit the form (although google automatically searches now without submitting)
+    # submit the form
     submit = driver.find_element_by_xpath("//a[@href='javascript:PC_7_0G3UNU10SD0MHTI7TQA0000000000000_validate()']")
     submit.click()
 
     # now we are on the password input page
     firstPassword = driver.find_element_by_name("memorableAnswer")
-    firstPassword.send_keys("some_password")
+    firstPassword.send_keys("your_first_pwd")
 
     # find all secondary password input
-    all_secondPassword = []
     all_secondPassword = driver.find_elements_by_xpath("//input[@type='password']")
 
-    valid_sec_pwd = []
-    i = 0
     # fill the second password
     for secondPwd in all_secondPassword:
         isDisable = secondPwd.get_attribute('disabled')
@@ -124,88 +127,129 @@ def download(_type, _month_year, _day ):
     logon = driver.find_element_by_xpath("//input[@class='submit_input']")
     logon.click()
 
-    # route to Card tab page
-    driver.get("https://www1.personal.ebanking.hsbc.com.hk/1/3/cards?__cmd-All_MenuRefresh=")
-
-    # click e-Statement hyperlink
-    link = driver.find_element_by_xpath("//a[@href='https://www1.personal.ebanking.hsbc.com.hk/1/3/my-hsbc/estatement?cmd-ECorrespDummyPortlet_in=&designateType=Card']")
-    link.click()
-
-    # find current window handle
-    global cur_win
-    cur_win = driver.window_handles[0]
-
-    # find month statement download link
-    all_mon_odd_rows = driver.find_elements_by_xpath("//tr[@class=' rowodd']")
-    all_mon_eve_rows = driver.find_elements_by_xpath("//tr[@class='zebra']")
-
-    global download_type
-    download_type = "CrCard"
-    for mon_odd_row in all_mon_odd_rows:
-        if done:
-            break
-        clickDownload(mon_odd_row, driver)
-
-    if type != 'd':
-        for mon_eve_row in all_mon_eve_rows:
-            if done:
-                break
-            clickDownload(mon_eve_row, driver)
-
     e.set()
-    print('\n' + "Credit card e-Statement download finished >>>>>>>")
-
-    # change the day to your saving account statement issue day
-    if _day < 23 and _type == 'd':
-        print("Current monthly saving account estatement is not issued.")
-        return
+    print('\n' + "Logged in to System !")
 
     e.clear()
-    t2 = threading.Thread(name='scheduler2',
-                          target=thread,
-                          args=(e, 1))
 
+    global cur_win
+    cur_win = ''
+    if 'c' in _crdr_type:
+        t3 = threading.Thread(name='scheduler3',
+                              target=thread,
+                              args=(e, 1))
+        t3.start()
 
-    t2.start()
-    # route to bank statement download page
-    done = False
-    all_ext_elements = driver.find_elements_by_xpath("//div[@class='extBgFix']")
+        # route to Card tab page
+        driver.get("https://www1.personal.ebanking.hsbc.com.hk/1/3/cards?__cmd-All_MenuRefresh=")
 
-    for ext_element in all_ext_elements:
-        link = ext_element.find_element_by_tag_name('a')
-        if "eStatement" in link.text:
-            link.click()
-            break
+        # sleep to wait until ad close
+        sleep(5)
 
-    # select 5xx account
-    select = Select(driver.find_element_by_xpath("//select[@id='number2']"))
-    select.select_by_value('account_no')
+        # click e-Statement hyperlink
+        link = driver.find_element_by_xpath("//a[@href='https://www1.personal.ebanking.hsbc.com.hk/1/3/my-hsbc/estatement?cmd-ECorrespDummyPortlet_in=&designateType=Card']")
+        link.click()
 
-    # click go
-    go_link = driver.find_element_by_xpath("//img[@src='/1/PA_defaultName/images/go.gif']")
-    go_link.click()
+        # find current window handle
+        cur_win = driver.window_handles[0]
 
-    # find current window handle
-    # global cur_win
-    cur_win = driver.window_handles[0]
+        # find month statement download link
+        all_mon_odd_rows = driver.find_elements_by_xpath("//tr[@class=' rowodd']")
+        all_mon_eve_rows = driver.find_elements_by_xpath("//tr[@class='zebra']")
 
-    # find month statement download link
-    all_mon_odd_rows = driver.find_elements_by_xpath("//tr[@class=' rowodd']")
-    all_mon_eve_rows = driver.find_elements_by_xpath("//tr[@class='zebra']")
-
-    # find month statement download link
-    # global download_type
-    download_type = "SavingAcc"
-    for mon_odd_row in all_mon_odd_rows:
-        if done:
-            break
-        clickDownload(mon_odd_row, driver)
-
-    if type != 'd':
-        for mon_eve_row in all_mon_eve_rows:
+        global download_type
+        download_type = "CrCard"
+        for mon_odd_row in all_mon_odd_rows:
             if done:
                 break
-            clickDownload(mon_eve_row, driver)
+            clickDownload(mon_odd_row, driver)
+            driver.switch_to.window(cur_win)
 
-    e.set()
-    print('\n' + 'Saving Account e-Statement download finished!>>>>>>>>>>>>>>>>')
+        if type != 'd':
+            for mon_eve_row in all_mon_eve_rows:
+                if done:
+                    break
+                clickDownload(mon_eve_row, driver)
+                driver.switch_to.window(cur_win)
+
+        e.set()
+        print('\n' + "Credit card e-Statement download finished !")
+
+    if 'd' in _crdr_type:
+        # begin to download saving account statements
+        if _day < 23 and _type == 'd':
+            print("Current monthly saving account estatement is not issued.")
+            return
+
+        e.clear()
+        t2 = threading.Thread(name='scheduler2',
+                              target=thread,
+                              args=(e, 1))
+        t2.start()
+
+        # route to Banking tab page
+        if _crdr_type == 'cd':
+            link = driver.find_element_by_xpath("//a[@id='Banking']")
+            link.click()
+        else:
+            driver.get("https://www2.personal.ebanking.hsbc.com.hk/1/2/banking?__cmd-All_MenuRefresh=")
+
+            # sleep to wait until ad close
+            sleep(5)
+
+            # click e-Statement hyperlink
+            link = driver.find_element_by_xpath(
+                "//a[@href='https://www2.personal.ebanking.hsbc.com.hk/1/2/my-hsbc/estatement?cmd-ECorrespDummyPortlet_in=&designateType=Banking']")
+            link.click()
+
+        # find current window handle
+        cur_win = driver.window_handles[0]
+
+        # route to bank statement download page
+        done = False
+
+        #banking_tab = driver.find_element_by_xpath("//a[@id='Banking']")
+        #banking_tab.click()
+        # all_ext_elements = driver.find_elements_by_xpath("//div[@class='extBgFix']")
+        #
+        # for ext_element in all_ext_elements:
+        #     link = ext_element.find_element_by_tag_name('a')
+        #     if "eStatement" in link.text:
+        #         link.click()
+        #         break
+
+        # select 5xx account
+        select = Select(driver.find_element_by_xpath("//select[@id='number2']"))
+        select.select_by_value('your_card_no')
+
+        # click go
+        go_link = driver.find_element_by_xpath("//img[@src='/1/PA_defaultName/images/go.gif']")
+        go_link.click()
+
+        # find current window handle
+        cur_win = driver.window_handles[0]
+
+        # find month statement download link
+        all_mon_odd_rows = driver.find_elements_by_xpath("//tr[@class=' rowodd']")
+        all_mon_eve_rows = driver.find_elements_by_xpath("//tr[@class='zebra']")
+
+        # find month statement download link
+        # global download_type
+        download_type = "SavingAcc"
+        for mon_odd_row in all_mon_odd_rows:
+            if done:
+                break
+            clickDownload(mon_odd_row, driver)
+            driver.switch_to.window(cur_win)
+
+        if type != 'd':
+            for mon_eve_row in all_mon_eve_rows:
+                if done:
+                    break
+                clickDownload(mon_eve_row, driver)
+                driver.switch_to.window(cur_win)
+
+        e.set()
+        print('\n' + 'Saving Account e-Statement download finished!')
+
+    #driver.close()
